finished writing decoded LED
