+++ core/signalprocessing/motionsenseHRVdecode/metadata/decoded_hrv_right_wrist.json
@@ -1,0 +1,40 @@
+{
+  "annotations":[],
+  "data_descriptor":[
+    {
+      "type":"array of Accelerometer,Gyroscope,LED",
+      "unit":"array"
+    }
+  ],
+  "execution_context":{
+    "processing_module":{
+      "name":"core.signalprocessing.decode",
+      "input_streams":[
+        {
+          "id":"CC_INPUT_STREAM_ID_CC",
+          "name":"CC_INPUT_STREAM_NAME_CC"
+        }
+      ],
+      "input_parameters":{
+        "sampling frequency": "25"
+      }
+    },
+    "algorithm":{
+      "method":"core.",
+      "authors":[
+        {
+          "name":"Md Azim Ullah",
+          "email":"mullah@memphis.edu"
+        }
+      ],
+      "version":"0.0.1",
+      "reference":{
+      },
+      "description":"Decodes the LED values from raw byte array and interpolates the timestamps based on sequence number"
+    }
+  },
+  "identifier":"CC_OUTPUT_STREAM_IDENTIFIER_CC",
+  "name":"org.md2k.signalprocessing.decodedLED.rightwrist",
+  "owner":"CC_OWNER_CC",
+  "type":"datastream"
+}
